# docs/swagger.yaml
definitions:
...
  status.Response:
    properties:
      status_code:
        type: integer
      status_message:
        type: string
    type: object
  status.ResponseWtihData:
    properties:
      data:
        type: any
      status:
        $ref: '#/definitions/status.Response'
    type: object
info:
  contact:
    email: service@invar.finance
    name: InVar
    url: https://www.invar.finance/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /admin/change_password_by_admin:
    post:
      consumes:
      - application/json
      description: 修改使用者密碼
      parameters:
      - description: 用戶ID
        in: body
        name: user_id
        required: true
        schema:
          type: string
      - description: 新密碼
        in: body
        name: password
        required: true
        schema:
          type: string
      - description: 確認新密碼
        in: body
        name: check_password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      security:
      - BearerAuth: []
      summary: 修改使用者密碼
      tags:
      - Auth
...
  /auth:
    post:
      consumes:
      - application/json
      description: 輸入帳號密碼與二階段驗證進行登入
      parameters:
      - description: 帳號
        in: body
        name: account
        required: true
        schema:
          type: string
      - description: 密碼
        in: body
        name: password
        required: true
        schema:
          type: string
      - description: 二階段驗證碼
        in: body
        name: tfa
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      summary: 使用者帳號登入
      tags:
      - Auth
...
  /change_password:
    post:
      consumes:
      - application/json
      description: 修改密碼
      parameters:
      - description: 舊密碼
        in: body
        name: old_password
        required: true
        schema:
          type: string
      - description: 新密碼
        in: body
        name: password
        required: true
        schema:
          type: string
      - description: 確認新密碼
        in: body
        name: check_password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      security:
      - BearerAuth: []
      summary: 修改密碼
      tags:
      - Auth
  /forget_password:
    post:
      consumes:
      - application/json
      description: 發送忘記密碼信
      parameters:
      - description: 信箱
        in: body
        name: email
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/status.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      summary: 發送忘記密碼信
      tags:
      - Auth
  /forget_password_access:
    post:
      consumes:
      - application/json
      description: 獲得忘記密碼令牌
      parameters:
      - description: 信箱
        in: body
        name: email
        required: true
        schema:
          type: string
      - description: 請求碼
        in: body
        name: request_token
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      summary: 獲得忘記密碼令牌
      tags:
      - Auth
...
  /register:
    post:
      consumes:
      - application/json
      description: 註冊新帳號
      parameters:
      - description: 名稱
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: 信箱
        in: body
        name: email
        required: true
        schema:
          type: string
      - description: 信箱驗證碼
        in: body
        name: email_code
        required: true
        schema:
          type: string
      - description: 密碼
        in: body
        name: password
        required: true
        schema:
          type: string
      - description: 密碼確認
        in: body
        name: password_check
        required: true
        schema:
          type: string
      - description: 邀請碼
        in: body
        name: referrer_code
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/status.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      summary: 註冊新帳號
      tags:
      - Auth
  /reset_password_with_token:
    post:
      consumes:
      - application/json
      description: 發送重設密碼
      parameters:
      - description: 信箱
        in: body
        name: email
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      summary: 發送重設密碼
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"